move api class in separate file
